# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IlliniHunt API",
    "version": "1.0.0",
    "summary": "Read and write the IlliniHunt catalog of University of Illinois Urbana-Champaign student, faculty, and staff projects.",
    "description": "IlliniHunt is a Product Hunt style launch board for the University of Illinois Urbana-Champaign community. This API exposes the public project catalog, categories, trending rankings, site statistics, and public user profiles without authentication, plus authenticated endpoints for submitting projects, voting, commenting, bookmarking, and managing collections.\n\nAll responses are JSON. Errors return a JSON object with an `error` field and, for unmatched paths, the requested `path`. The API is rate limited to 300 requests per minute per client (see the `RateLimit` response headers, RFC 9331 draft-7 format).\n\nAuthenticated endpoints require a Microsoft Entra ID bearer token issued for the University of Illinois tenant. Sign-in is restricted to @illinois.edu accounts; there is no public self-service API key programme.",
    "contact": {
      "name": "IlliniHunt maintainers",
      "url": "https://github.com/gies-ai-experiments/illinihunt/issues"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://illinihunt.azurewebsites.net",
      "description": "Production API (Azure App Service)"
    }
  ],
...
  "paths": {
...
    "/api/users/{id}": {
      "get": {
        "operationId": "getUserProfile",
        "tags": [
          "Users"
        ],
        "summary": "Get a public user profile",
        "description": "Returns the public profile fields for one user by UUID. No authentication. Email addresses are not returned.",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "responses": {
          "200": {
            "description": "The profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "user"
                  ],
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/UserProfile"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/users/{id}/projects": {
      "get": {
        "operationId": "listUserProjects",
        "tags": [
          "Users"
        ],
        "summary": "List a user's projects",
        "description": "Returns the active projects submitted by one user. No authentication.",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "responses": {
          "200": {
            "description": "The user's projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "entraBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Microsoft Entra ID access token for the University of Illinois tenant. Obtained interactively through the IlliniHunt web app; sign-in is restricted to @illinois.edu accounts. There is no public API key issuance, so agents can use the unauthenticated read endpoints only."
      }
    },
    "parameters": {
...
      "UserId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "User UUID.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "No such resource",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "description": "Every error response is JSON with a human-readable error message.",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "path": {
            "type": "string",
            "description": "Requested path, present on unmatched-route 404s"
          }
        }
      },
      "Category": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "examples": [
              "Emerging Technology"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Icon name used by the web UI"
          },
          "color": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hex colour used by the web UI"
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "UserSummary": {
        "type": "object",
        "required": [
          "id"
        ],
        "description": "Public profile fields. Email addresses are never included.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          },
          "full_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatar_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          }
        }
      },
      "Project": {
        "type": "object",
        "required": [
          "id",
          "name",
          "tagline"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "examples": [
              "CampusCloset"
            ]
          },
          "tagline": {
            "type": "string",
            "description": "One-line pitch shown on cards"
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Long-form description"
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "website_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "github_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "category_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "upvotes_count": {
            "type": [
              "integer",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ],
            "examples": [
              "active"
            ]
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "categories": {
            "$ref": "#/components/schemas/Category"
          },
          "users": {
            "$ref": "#/components/schemas/UserSummary"
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "required": [
          "id"
        ],
        "description": "Public profile. Email addresses are deliberately not exposed on unauthenticated endpoints.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          },
          "full_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatar_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "github_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "linkedin_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "website_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "year_of_study": {
            "type": [
              "string",
              "null"
            ]
          },
          "department": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_verified": {
            "type": "boolean"
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      }
    }
  },
  "security": []
}